drain-to-feat for vehicle-imports — relocate Autosys import UI
components/external-inputs/autosys/ was the entire UI for the vehicle-imports
feature, but data/vehicle-imports/ held only the data layer. Per README's
"where new files go" rule, entity-specific UI belongs in data/<feature>/.
Move the MultiImport stepper (5 files) and the launcher FAB into
data/vehicle-imports/components/; the FAB also leaves the generic
components/dialogs/ sub-area.

Delete the dead SingleImport/SingleImportQuery/SingleImportConfirm trio —
SingleImport was deactivated (posts raw Autosys XML without injecting
OperationalNumber) and had no importers.

Drop external-inputs/ from the README components/ sub-area list; add a
FORK_DECISIONS.md ADR recording the move and the components/-subdir choice.
