+++ src/data/vehicle-imports/components/AutosysImportFloatingMenu.tsx
@@ -3,7 +3,7 @@ import { Dialog, Fab, Tooltip } from '@mui/material';
 import { useState } from 'react';
 import { useTranslation } from 'react-i18next';
 import { useNavigate } from 'react-router-dom';
-import MultiImport from '../external-inputs/autosys/MultiImport';
+import MultiImport from './MultiImport';
 
 export default function AutosysImportFloatingMenu() {
   const [open, setOpen] = useState(false);
