+++ src/data/vehicle-imports/components/MultiImportColumnMapper.tsx
@@ -12,7 +12,7 @@ import TableContainer from '@mui/material/TableContainer';
 import TableHead from '@mui/material/TableHead';
 import TableRow from '@mui/material/TableRow';
 import Typography from '@mui/material/Typography';
-import type { TableMeta } from '../../../data/vehicle-imports/inputTextAnalyzer';
+import type { TableMeta } from '../inputTextAnalyzer';
 
 const NONE = '__none__';
 const PREVIEW_ROWS = 5;
