+++ src/data/vehicle-imports/components/MultiImportFileInput.tsx
@@ -3,12 +3,8 @@ import { useTranslation } from 'react-i18next';
 import { CloudUpload } from '@mui/icons-material';
 import Box from '@mui/material/Box';
 import Typography from '@mui/material/Typography';
-import {
-  type AnalyzerResult,
-  type TableMeta,
-  inputTextAnalyzer,
-} from '../../../data/vehicle-imports/inputTextAnalyzer';
-import type { RegNumbersStatus } from '../../../data/vehicle-imports/regNumbersTextTransformer';
+import { type AnalyzerResult, type TableMeta, inputTextAnalyzer } from '../inputTextAnalyzer';
+import type { RegNumbersStatus } from '../regNumbersTextTransformer';
 
 interface MultiImportFileInputProps {
   onParsed: (regNumbers: string[], status: RegNumbersStatus) => void;
