+++ src/data/vehicle-types/VehicleTypeView.tsx
@@ -2,7 +2,7 @@ import { vehicleTypeViewConfig } from './vehicleTypeViewConfig.tsx';
 import GenericDataViewPage from '../../pages/GenericDataViewPage.tsx';
 import type { VehicleType } from './vehicleTypeTypes.ts';
 import type { OrderBy } from './useVehicleTypes.ts';
-import AutosysImportFloatingMenu from '../../components/dialogs/AutosysImportFloatingMenu.tsx';
+import AutosysImportFloatingMenu from '../vehicle-imports/components/AutosysImportFloatingMenu.tsx';
 import { useUrlFilters } from '../../hooks/useUrlFilters.ts';
 
 export default function VehicleTypeView() {
